requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
autopep8==2.3.2
click==8.2.1
colorama==0.4.6
databases==0.9.0
fastapi==0.115.12
greenlet==3.2.2
h11==0.16.0
idna==3.10
psycopg2==2.9.10
pycodestyle==2.13.0
pydantic==2.11.5
pydantic_core==2.33.2
python-dotenv==1.1.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.13.2
uvicorn==0.34.2
